Move back changes to original image operators files
